// frontend/src/app/core/api/model/period-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PeriodRequest { 
    label: string;
    startDate: string;
    endDate: string;
    academicYearId: number;
}